+++ lib/modules/test/test_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: test
-// Updated at: 2026-04-24 21:30:40
-// Commit: feat(test): add bi-directional event bus handler for test
+// Updated at: 2026-04-24 21:54:09
+// Commit: perf(test): optimize memory consumption in test list viewport
 
 class testComponentState {
-  final String commitRef = '649';
-  final String lastMessage = 'feat(test): add bi-directional event bus handler for test';
-  final DateTime timestamp = DateTime.parse('2026-04-24T21:30:40+05:00');
+  final String commitRef = '650';
+  final String lastMessage = 'perf(test): optimize memory consumption in test list viewport';
+  final DateTime timestamp = DateTime.parse('2026-04-24T21:54:09+05:00');
 }